move the legal documents to the legal-notices endpoint
