+++ pages/legal-notices/cookie-policy.js
